# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A social network API built with Go. Supports user authentication, posts, comments, reactions, groups, messaging, notifications, and real-time WebSocket events.",
        "title": "Social Network API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/me": {
            "get": {
                "description": "Returns profile data for the currently authenticated user based on session cookie.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "Authenticated user data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/online-users": {
            "get": {
                "description": "Returns the list of user IDs currently connected via WebSocket.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WebSocket"
                ],
                "summary": "Get online users",
                "responses": {
                    "200": {
                        "description": "List of online user IDs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/session/validate": {
            "get": {
                "description": "Lightweight endpoint to check if the session cookie is still valid. Used by NextAuth for server-side session validation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Validate session",
                "responses": {
                    "200": {
                        "description": "Session is valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Invalid or expired session",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/search": {
            "get": {
                "description": "Searches for users by firstname, lastname, or nickname.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query string",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users fetched",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.userSearchResult"
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "description": "Returns basic user info (id, nickname, firstname, lastname, avatar, is_private).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data fetched",
                        "schema": {
                            "$ref": "#/definitions/handlers.userProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/ws": {
            "get": {
                "description": "Upgrades an HTTP connection to a WebSocket connection after ticket verification. The ticket is obtained from /api/ws-ticket.",
                "tags": [
                    "WebSocket"
                ],
                "summary": "WebSocket connection handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "One-time WebSocket ticket",
                        "name": "ticket",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid or missing ticket",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.userProfileResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_private": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "handlers.userSearchResult": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_private": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}